refactor (remove) misc_utils
