+++ src/fprime_gds/common/gds_cli/command_send.py
@@ -118,7 +118,36 @@ class CommandSendCommand(BaseCommand):
         :param json: Whether or not to return a JSON representation of "temp"
         :return: A readable string version of "item"
         """
-        return misc_utils.get_cmd_template_string(item, json)
+        if not item:
+            return ""
+        if json:
+            # This is questionable whether it should be supported
+            from fprime_gds.flask.json import getter_based_json
+            return str(getter_based_json(item))
+
+        cmd_string = "%s (%d) | Takes %d arguments.\n" % (
+            item.get_full_name(),
+            item.get_id(),
+            len(item.get_args()),
+        )
+
+        cmd_description = item.get_description()
+        if cmd_description:
+            cmd_string += f"Description: {(cmd_description)}\n" 
+
+        for arg in item.get_args():
+            arg_name, arg_description, arg_type = arg
+            if not arg_description:
+                arg_description = "--no description--"
+            # Can't compare against actual module, since EnumType is a serializable
+            # type from the dictionary
+            if type(arg_type).__name__ == "EnumType":
+                arg_description = f"{str(arg_type.keys())} | {arg_description} "
+
+            cmd_string += f"\t{arg_name} ({arg_type.__name__}): {arg_description}\n"
+
+        return cmd_string
+
 
     @classmethod
     def _execute_command(cls, args, api: IntegrationTestAPI):
